refactor(DEV-11087): move counterpart cell to common dir to reuse
