+++ packages/sdk-react/src/ui/CounterpartCell/CounterpartCell.tsx
