+++ packages/sdk-react/src/ui/CounterpartCell/index.ts
